Fix Docker/Brownie transaction EventDict issue
Was using Brownie `Chain()` to retrieve transaction and
`tx.events["Transfer"]` to access ERC20 transfer events. For some
reason the transaction events were decoded and keyed properly when run
on my local machine, but returned as raw topics when run in a Docker
container.

To address, used web3py instead to grab the transaction receipt and
logs, which were then decoded using the Transfer event ABI.